Final painter program (base)
Migrating the project from Google Drive to GitHub as the base is complete. Fixed and refactored the painting program code and created a quick demo on Colab. Refactored content to match project template. Will continue to store a copy of the project on Google Drive for backup, and to store the checkpoints for the networks.