Added evaluation function
Added a function to evaluate the paintings and store the results in csv documents.

It calculates mse, mrse, l1 distance, peak signal to noise ratio and stuctural similarity between the generated and the target image.

It also performs the calculations only on the most important parts of the images filtering with the map.